invoice_platform/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.8
django-filter==23.5
django-seed==0.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==22.6.0
idna==3.6
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
toposort==1.10
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
